ADDED - Invasion and Deploy form components
